Feature/layout (#3)
* Refactor UI and integrate camera functionality

Refactored layout to enhance modularity, including the addition of reusable header, footer, and demo components. Introduced a `CameraCapture` component to enable capturing images directly from the camera. Reorganized and improved the overall page structure and styling for a better user experience.

* Rename dynamic route parameter from 'id' to 'image'

Updated the dynamic route in the image module to use 'image' as the parameter instead of 'id'. This improves clarity and consistency with the naming conventions used across the codebase.

* Enhance image processing functionality and improve button interactivity

* Add image processing features and enhance UI with loading states and skeleton cards